LayoutTests/http/tests/resourceLoadStatistics/third-party-cookie-blocking-ephemeral-expected.txt

Tests that all third-party cookies are blocked (ephemeral session).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Resource load statistics is running in ephemeral mode.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Before user interaction, should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame2-->'
--------
After user interaction, should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Client-side document.cookie:
